Added style and AJAX logic to compile DSL
