+++ src/main/java/org/dsl/geometry/compiler/controllers/WebController.java
@@ -1,0 +1,12 @@
+package org.dsl.geometry.compiler.controllers;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class WebController {
+  @GetMapping("/")
+  public String index() {
+    return "index";
+  }
+}
