--- src/main/resources/templates/result.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>Результаты Компиляции</title>
-</head>
-<body>
-<h1>Результаты Компиляции</h1>
-<img th:src="@{${imagePath}}" />
-<a href="/">Вернуться к вводу кода</a>
-</body>
-</html>
